+++ src/main/php/lang/ast/emit/PHP70.class.php
@@ -1,7 +1,7 @@
 <?php namespace lang\ast\emit;
 
 use lang\ast\Node;
-use lang\ast\nodes\{Assignment, InstanceExpression, ScopeExpression, Literal, Variable};
+use lang\ast\nodes\{InstanceExpression, ScopeExpression, Literal, Variable};
 use lang\ast\types\{IsUnion, IsIntersection, IsFunction, IsArray, IsMap, IsNullable, IsValue, IsLiteral, IsGeneric};
 
 /**
@@ -14,6 +14,7 @@ class PHP70 extends PHP {
     ArbitrayNewExpressions,
     ArrayUnpackUsingMerge,
     AttributesAsComments,
+    ForeachDestructuringAsStatement,
     MatchAsTernaries,
     NullsafeAsTernaries,
     OmitArgumentNames,
@@ -129,30 +130,5 @@ class PHP70 extends PHP {
     }
 
     return parent::emitAssignment($result, $assignment);
-  }
-
-  protected function emitForeach($result, $foreach) {
-    $result->out->write('foreach (');
-    $this->emitOne($result, $foreach->expression);
-    $result->out->write(' as ');
-    if ($foreach->key) {
-      $this->emitOne($result, $foreach->key);
-      $result->out->write(' => ');
-    }
-
-    // Rewrite destructuring assignment as first statement in loop
-    if ('array' === $foreach->value->kind) {
-      $t= $result->temp();
-      $result->out->write($t.') {');
-      $this->rewriteDestructuring($result, new Assignment($foreach->value, '=', new Variable(substr($t, 1))));
-      $result->out->write(';');
-      $this->emitAll($result, $foreach->body);
-      $result->out->write('}');
-    } else {
-      $this->emitOne($result, $foreach->value);
-      $result->out->write(') {');
-      $this->emitAll($result, $foreach->body);
-      $result->out->write('}');
-    }
   }
 }