+++ ChangeLog.md
@@ -3,6 +3,8 @@ XP Compiler ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Fixed issue #31: Support `list()` reference assignment for PHP < 7.3
+  (@thekid)
 * Fixed `foreach` statement using destructuring assignments in PHP 7.0
   (@thekid)
 * Merged PR #146: Add support for omitting expressions in destructuring